HW2
The second homework where 2 main functinos where created. 1)  Read() which reads a file and returns the total time, the total amount of particles, and a numpy array compramising of the rest of the data. 2) ParticleInfo() which computes the mass of a certain particle under the particle type, the magnitude of all the distances and magnitude of all the velocities