HW3
The third homework compromising of 5 files. Two of them are the python files named ReadFile which reads a specific file, GalaxyMass which returns the total mass of any galaxy for the specific component of that particle type. Question4.docx is the answers contained for the questions in problem 4. Datatable.tex is the latex file used to produce the table seen in DataTable.pdf