HW4
This is the 4th homework with 4 files. ReadFile is a file used to read data and used in CenterOfmass.py. This function computes the center of mass position and mass and velocity by using classes. PrintLine.png is the print statement that proves the results while Answer.txt is the same expect in txt format.